requirements.txt
flake8==5.0.4
flake8-docstrings==1.6.0
python-dotenv==0.19.0
python-telegram-bot==13.4